Extract unification code to unify module
